+++ tests/Transformer/NoSchemaTransformer.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tomaj\NetteApi\Test\Transformer;
+
+use League\Fractal\TransformerAbstract;
+
+class NoSchemaTransformer extends TransformerAbstract
+{
+}